FIX BULLSHIT FUCKING DAMAGEABLESYSTEM WARZONE
