+++ Content.Shared/Damage/Components/DamageableComponent.Goob.cs
@@ -1,0 +1,7 @@
+namespace Content.Shared.Damage;
+
+public sealed partial class DamageableComponent
+{
+    [ViewVariables]
+    public TimeSpan LastModifiedTime = TimeSpan.Zero;
+}
